requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
tensorflow-cpu==2.13.0
Pillow>=9.5.0
numpy>=1.23.0
gdown>=4.7.1
